Get visualize output to populate
Unfortunately, I was unable to get the graphs to process within the RunAllScripts and pass them to testPlot() to just read in and populate the 'Visualize Outputs' page. Came up with another solution to get a variable amount of graphics onto the page automatically, without needing more user inputs by filling an empty list of 6 and using if statements to only pull the number of climate models selected from the list.